Refactor static language generator helpers
