+++ tablegen/src/static_language/mod.rs
@@ -1,0 +1,6 @@
+//! Single-responsibility helpers for [`StaticLanguageGenerator`].
+
+pub(crate) mod compression_tokens;
+pub(crate) mod node_types;
+pub(crate) mod symbols;
+pub(crate) mod table_tokens;
